+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - Label component
 - Label storybook
+- Fieldset component
+- Fieldset storybook
 
 ### Changed